requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.7.0
h11==0.14.0
idna==3.10
multidict==6.6.4
outcome==1.3.0.post0
propcache==0.3.2
py-cord==2.6.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.3
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
